common: annotate in which dCache version Principals were added
Motivation:

When adding a Principal the gPlazma door must filter out the new
principal.  This is because we guarantee compatibility with older
components; in particular, with older pool nodes.  Without filtering out
the principal, the Principal can travel to the pool (as part of a
Subject), which will then not be able to deserialise the message.

It is also important to know when it is safe to remove the Principal
filtering.

Therefore, it makes sense to record when a Principal was added to
dCache.

Modification:

Add @since annotation for Principals.

Result:

No user- or admin obserable changes.

Target: master
Requires-notes: no
Requires-book: no
Patch: https://rb.dcache.org/r/12340/
Acked-by: Tigran Mkrtchyan
Acked-by: Lea Morschel